add merge operator and reimplement incref and decref

+++ src/database/store/entry.rs
@@ -7,6 +7,6 @@ pub(crate) struct Entry<Key: Field, Value: Field> {
 
     #[serde(bound(deserialize = "Node<Key, Value>: Deserialize<'de>"))]
     pub node: Node<Key, Value>,
-    pub references: usize,
+    pub references: i32,
 }
